update mutation and adresse entities

+++ src/main/java/com/apeiron/immoxperts/repository/LocalRepository.java
@@ -9,4 +9,4 @@ import org.springframework.stereotype.Repository;
  */
 @SuppressWarnings("unused")
 @Repository
-public interface LocalRepository extends JpaRepository<Local, Long> {}
+public interface LocalRepository extends JpaRepository<Local, Integer> {}
