+++ src/main/java/com/apeiron/immoxperts/repository/LotRepository.java
@@ -9,4 +9,4 @@ import org.springframework.stereotype.Repository;
  */
 @SuppressWarnings("unused")
 @Repository
-public interface LotRepository extends JpaRepository<Lot, Long> {}
+public interface LotRepository extends JpaRepository<Lot, Integer> {}
